+++ Hillbilly Skeet Shoot/Assets/Scripts/GameManagerScript.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.SceneManagement;
+
+public class GameManagerScript : MonoBehaviour
+{
+    public static int score;
+    public Text scoreText;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        scoreText.text = "Score: " + score;
+    }
+}
