+++ Hillbilly Skeet Shoot/Assets/Scripts/GameManagerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e46614608bead0148bdf5f56a2b63cde
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
